scrapers/api/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
pydantic-settings==2.0.3
pydantic[email]==2.3.0
motor==3.3.1
pymongo==4.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.0.0
python-dotenv==1.0.0
aiofiles==23.2.1
jinja2==3.1.2
httpx==0.25.0
tenacity==8.2.3
aiocache==0.12.1
aiohttp==3.8.5
aiosmtplib==2.0.2
pillow==10.0.1
python-slugify==8.0.1
pytz==2023.3
uuid==1.30
ujson==5.8.0
orjson==3.9.7
rich==13.5.3
loguru==0.7.2 